Let go of dead code
